Adding support for term duplicate check on the committee ID field on the taxonomy terms.
The solution is backward compatible and is not locked to the committee
vocabulary. “$term_name” can be provided as an array with the keys
esdh_name (The term name), esdh_id (the term ID from the ESDH system)
and esdh_id_field (the field machine name used to store the ID on the
taxonomy term e.g. field_os2web_meetings_commiteeid).

The FICS integration is using this to avoid problems with committees
with the same name, and to allow automatically renaming of committees
directly from FICS.